backend/requirements.txt
Brotli==1.1.0
bs4==0.0.2
cloudscraper==1.2.71
exa-py==1.8.8
fastapi==0.115.5
faiss-cpu==1.9.0
feedparser==6.0.11
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
langgraph==0.2.18
langsmith==0.1.114
python-jose==3.3.0
pdfplumber
python-dotenv
redis==5.2.1
urllib3
uvicorn
websockets==14.1
